More refactoring for run
